refactor(web): split admin categories into form + list, de-carded
